nvme: Clear the notify flag if the consumer rejects the controller.

+++ usr/src/contrib/bhyve/dev/nvme/nvme.h
@@ -2072,7 +2072,7 @@ void	nvme_resv_status_swapbytes(struct nvme_resv_status *s __unused,
     size_t size __unused)
 {
 #ifndef _LITTLE_ENDIAN
-	u_int i, n;
+	size_t i, n;
 
 	s->gen = le32toh(s->gen);
 	n = (s->regctl[1] << 8) | s->regctl[0];
@@ -2090,7 +2090,7 @@ void	nvme_resv_status_ext_swapbytes(struct nvme_resv_status_ext *s __unused,
     size_t size __unused)
 {
 #ifndef _LITTLE_ENDIAN
-	u_int i, n;
+	size_t i, n;
 
 	s->gen = le32toh(s->gen);
 	n = (s->regctl[1] << 8) | s->regctl[0];
